+++ tests/at_end2end_test/test/notify_test.dart
@@ -3,7 +3,6 @@ import 'dart:convert';
 import 'package:at_client/at_client.dart';
 import 'package:at_end2end_test/config/config_util.dart';
 import 'package:at_end2end_test/src/test_initializers.dart';
-import 'package:at_end2end_test/src/test_preferences.dart';
 import 'package:at_end2end_test/utils/test_constants.dart';
 import 'package:test/test.dart';
 import 'package:uuid/uuid.dart';
@@ -41,9 +40,8 @@ void main() async {
     // for each run.
     var value = '+1 100 200 30';
     // Setting currentAtSign atClient instance to context.
-    currentAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(currentAtSign, namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
+    currentAtClientManager = await TestSuiteInitializer.getInstance()
+        .switchToAtSign(currentAtSign, namespace);
     final notificationResult = await currentAtClientManager
         .atClient.notificationService
         .notify(NotificationParams.forUpdate(phoneKey, value: value));
@@ -52,10 +50,8 @@ void main() async {
         NotificationStatusEnum.delivered);
 
     // Setting sharedWithAtSign atClient instance to context.
-    await AtClientManager.getInstance().setCurrentAtSign(
-        sharedWithAtSign,
-        namespace,
-        TestPreferences.getInstance().getPreference(sharedWithAtSign));
+    await TestSuiteInitializer.getInstance()
+        .switchToAtSign(sharedWithAtSign, namespace);
     var notificationListResult = await AtClientManager.getInstance()
         .atClient
         .notifyList(regex: 'phone$randomValue');
@@ -82,9 +78,8 @@ void main() async {
     // for each run.
     var value = '+1 100 200 30';
     // Setting currentAtSign atClient instance to context.
-    currentAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(currentAtSign, namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
+    currentAtClientManager = await TestSuiteInitializer.getInstance()
+        .switchToAtSign(currentAtSign, namespace);
     final notificationResult = await currentAtClientManager
         .atClient.notificationService
         .notify(NotificationParams.forUpdate(phoneKey, value: value),
@@ -94,10 +89,8 @@ void main() async {
         NotificationStatusEnum.delivered);
 
     // Setting sharedWithAtSign atClient instance to context.
-    await AtClientManager.getInstance().setCurrentAtSign(
-        sharedWithAtSign,
-        namespace,
-        TestPreferences.getInstance().getPreference(sharedWithAtSign));
+    await TestSuiteInitializer.getInstance()
+        .switchToAtSign(sharedWithAtSign, namespace);
     var notificationListResult = await AtClientManager.getInstance()
         .atClient
         .notifyList(regex: 'phone$randomValue');